feat: Add warning for partial data in OHS analysis
This commit adds a check to determine if all required months are present in the data points for the OHS analysis. If not, it returns a warning to the user indicating that only partial data is available, along with the calculated results. This improves the transparency and accuracy of the analysis by informing the user when the results might be based on incomplete data.